requirements.txt
streamlit==1.24.0
pdf2image==1.16.3
easyocr==1.6.2
numpy==1.24.0
torch==2.0.1
google-cloud-translate==2.0.1
langdetect==1.0.9
Pillow==9.4.0
